requirements.txt
Django==4.1.2
djangorestframework==3.13.1
djangorestframework-social-oauth2==1.0.0
django-allauth==0.51.0
django-cors-headers==3.13.0
requests==2.28.1
social-auth-app-django==5.2.0